removing logic from vcenter driver to command_orchestrator
